Create 2017-11-10_meeting_notes.md
These are the compiled notes from the meeting we had with Dr. Peter Caie. I have tried to sort the bullet points into relevant groups.
Bear in mind that this is the ideal description of the software and that it is up to us to figure out what is feasible in our groups and in the time we have to do it.